move callback and wait handling to middlewares
