+++ Makefile
@@ -6,7 +6,7 @@ TEST_TARGET ?= test/
 
 lint:
 	echo "  $(P) Linting"
-	$(BIN_DIR)/eslint lib && $(BIN_DIR)/eslint test && $(BIN_DIR)/eslint plugins
+	$(BIN_DIR)/eslint lib && $(BIN_DIR)/eslint test && $(BIN_DIR)/eslint middlewares
 
 test: lint
 	echo "  $(P) Testing"
