+++ lib/collection.js
@@ -1,25 +1,5 @@
-// var helpers = require('./helpers')
-// var debug = require('debug')('monk:queries')
 var objectAssign = require('object-assign')
 var applyMiddlewares = require('./applyMiddlewares')
-
-function thenFn (fn) {
-  return function (res) {
-    if (fn && typeof fn === 'function') {
-      fn(null, res)
-    }
-    return res
-  }
-}
-
-function catchFn (fn) {
-  return function (err) {
-    if (fn && typeof fn === 'function') {
-      return fn(err)
-    }
-    throw err
-  }
-}
 
 module.exports = Collection
 
@@ -32,21 +12,10 @@ var DEFAULT_OPTIONS = {
     require('monk-plugin-query'),
     require('monk-plugin-options'),
     require('monk-plugin-cast-ids')(FIELDS_TO_CAST),
-    require('monk-plugin-fields')
+    require('monk-plugin-fields'),
+    require('monk-plugin-handle-callback'),
+    require('monk-plugin-wait-for-connection')
   ]
-}
-
-function wrapperMiddleware (context) {
-  return function (next) {
-    return function (args) {
-      return context.monkInstance.executeWhenOpened().then(function (db) {
-        return db.collection(context.collection.name)
-      }).then(function (col) {
-        args.col = col
-        return next(args)
-      }).then(thenFn(args.callback)).catch(catchFn(args.callback))
-    }
-  }
 }
 
 function Collection (manager, name, options) {
@@ -54,7 +23,7 @@ function Collection (manager, name, options) {
   this.name = name
   this.options = objectAssign({}, DEFAULT_OPTIONS, options || {})
 
-  this.middlewares = (this.options.middlewares || []).concat(wrapperMiddleware)
+  this.middlewares = this.options.middlewares || []
   delete this.options.middlewares
 
   delete this.options.plugins
