+++ middlewares/handle-callback/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "monk-plugin-handle-callback",
+  "description": "A monk plugin to handle callback",
+  "version": "0.2.0",
+  "main": "index.js",
+  "keywords": [
+    "monk",
+    "plugin",
+    "callback"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/Automattic/monk.git"
+  },
+  "license": "MIT"
+}
