+++ middlewares/wait-for-connection/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "monk-plugin-wait-for-connection",
+  "description": "A monk plugin to wait for the connection",
+  "version": "0.2.0",
+  "main": "index.js",
+  "keywords": [
+    "monk",
+    "plugin",
+    "wait",
+    "connection"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/Automattic/monk.git"
+  },
+  "license": "MIT"
+}
